[misc][pat] renamed pathroutes.js to routes.js
