+++ src/components/SignupCard/index.jsx
@@ -5,7 +5,7 @@ import { Button, Card, Form, Input, Space, Typography } from 'antd';
 import { LockOutlined, UserOutlined, MailOutlined } from '@ant-design/icons';
 
 import { useAuth } from 'services/AuthProvider';
-import routes from 'constants/pathroutes';
+import routes from 'constants/routes';
 
 import './styles.scss';
